langgraph/requirements.txt
langgraph 
langsmith
langchain[openai]
ipykernel
python-dotenv
langchain_tavily
uuid
numpy